ecom_pipeline/producer/requirements.txt
kafka-python-ng>=2.2.0
requests==2.31.0
